解决订单管理中查看详情时无法显示图片的bug

+++ src/main/java/com/lsh/mustardtest/pojo/UserExample.java
@@ -384,7 +384,7 @@ public class UserExample {
             return (Criteria) this;
         }
 
-        public Criteria andPhoneEqualTo(Integer value) {
+        public Criteria andPhoneEqualTo(Long value) {
             addCriterion("phone =", value, "phone");
             return (Criteria) this;
         }
